Fixed: case of api failure as the endpoints using app's domain instead of oms domain(#504)
When making api call to fetch identification for products it gets failed as unauthorized as the token is not set till the api call. As the api fails, it clears the oms-url and as the apps flow does not break for this api failure some other apis gets called with the app's url instead of oms-url.

Moved the logic to update the token to handle the api failure case so that endpoints will not be prepared for app. Also added a check on app's mount that the identification method should only be called if we have a token available.